src/test/java/com/redhat/ConfigurationModelTest.java
package com.redhat;

import static org.junit.jupiter.api.Assertions.assertNotNull;

import java.io.File;
import java.io.IOException;

import org.junit.jupiter.api.Test;

import com.fasterxml.jackson.core.exc.StreamReadException;
import com.fasterxml.jackson.databind.DatabindException;

import io.quarkus.test.junit.QuarkusTest;

@QuarkusTest
public class ConfigurationModelTest {
    
    @Test
    public void testLoadYaml() throws StreamReadException, DatabindException, IOException {
        ConfigurationModel model1 = ConfigurationModel.loadFromFile(new File("src/test/resources/example1.yaml"));
        assertNotNull(model1);
        ConfigurationModel model2 = ConfigurationModel.loadFromFile(new File("src/test/resources/example2.yaml"));
        assertNotNull(model2);
    }
}
